+++ CHANGELOG.md
@@ -24,7 +24,12 @@ _the graphics engine_
 
 ---
 
-#### [0.5.4] pipelines (00 sept 22_)
+#### [0.5.5] attachments (00 sept 22_)
+
+- **added** - create vkimage and vkimageview
+- **added** - texture data struct and creation
+
+#### [0.5.4] pipelines (17 sept 22_)
 
 - **added** - pipeline layout
 - **changed** - moved r_vulkan and r_types contents to r_api and renamed r_vulkan_api to r_api_vulkan